+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
  - Using new logging context
+ - Rewrote RPC JSON serialization
 
 ## [12.0.0] - 2020-03-20
 ### Added